Refactor hooks structure
